Add DetailSectionView & divided property from DetailView
